requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.7.14
charset-normalizer==3.4.2
coloredlogs==15.0.1
facenet-pytorch==2.5.3
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
onnx==1.18.0
onnx-tf==1.10.0
onnxruntime==1.19.2
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
Pygments==2.19.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
torch==2.7.1
torchvision==0.22.1
typeguard==2.13.3
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0